Polished our design_document.md and accessibility.md. Move RegisterHelper from usecase to controllers so it does not violate Clean Architecture.
