+++ tcard/src/main/java/controllers/RegisterController.java
@@ -1,6 +1,5 @@
 package controllers;
 
-import usecases.RegisterUseCase;
 import interfaces.UserReadWriter;
 import java.util.*;
 
@@ -8,23 +7,23 @@ import java.util.*;
  * Controls the process for user registration.
  */
 public class RegisterController {
-    RegisterUseCase registerUseCase;
-    final Map<RegisterUseCase.RegistrationResult, String> REGISTER_MESSAGES = new EnumMap<>(RegisterUseCase.RegistrationResult.class);
+    RegisterHelper registerHelper;
+    final Map<RegisterHelper.RegistrationResult, String> REGISTER_MESSAGES = new EnumMap<>(RegisterHelper.RegistrationResult.class);
 
     /**
      * Constructor for RegisterController
      * @param userReadWriter an instance of the database helper interface UserReadWriter
      */
     public RegisterController(UserReadWriter userReadWriter) {
-        this.registerUseCase = new RegisterUseCase(userReadWriter);
+        this.registerHelper = new RegisterHelper(userReadWriter);
 
-        REGISTER_MESSAGES.put(RegisterUseCase.RegistrationResult.EMPTY_FIELD, "Please fill all the fields");
-        REGISTER_MESSAGES.put(RegisterUseCase.RegistrationResult.EMAIL_ERROR, "Email must be a U of T email");
-        REGISTER_MESSAGES.put(RegisterUseCase.RegistrationResult.TCARD_NUM_ERROR, "TCard number must be 10 digits long");
-        REGISTER_MESSAGES.put(RegisterUseCase.RegistrationResult.YEAR_ERROR, "Year must be greater than 0");
-        REGISTER_MESSAGES.put(RegisterUseCase.RegistrationResult.USER_EXISTS, "User already exists! Please sign in");
-        REGISTER_MESSAGES.put(RegisterUseCase.RegistrationResult.REGISTER_FAILURE, "Registration failed");
-        REGISTER_MESSAGES.put(RegisterUseCase.RegistrationResult.REGISTER_SUCCESS, "Registered successfully");
+        REGISTER_MESSAGES.put(RegisterHelper.RegistrationResult.EMPTY_FIELD, "Please fill all the fields");
+        REGISTER_MESSAGES.put(RegisterHelper.RegistrationResult.EMAIL_ERROR, "Email must be a U of T email");
+        REGISTER_MESSAGES.put(RegisterHelper.RegistrationResult.TCARD_NUM_ERROR, "TCard number must be 10 digits long");
+        REGISTER_MESSAGES.put(RegisterHelper.RegistrationResult.YEAR_ERROR, "Year must be greater than 0");
+        REGISTER_MESSAGES.put(RegisterHelper.RegistrationResult.USER_EXISTS, "User already exists! Please sign in");
+        REGISTER_MESSAGES.put(RegisterHelper.RegistrationResult.REGISTER_FAILURE, "Registration failed");
+        REGISTER_MESSAGES.put(RegisterHelper.RegistrationResult.REGISTER_SUCCESS, "Registered successfully");
     }
 
     /**
@@ -34,7 +33,7 @@ public class RegisterController {
      * @return a string with a message describing the status of the registration
      */
     public String runRegister(List<String> userRegistrationInfo) {
-        RegisterUseCase.RegistrationResult registerResult = registerUseCase.registerUser(userRegistrationInfo);
+        RegisterHelper.RegistrationResult registerResult = registerHelper.registerUser(userRegistrationInfo);
         return REGISTER_MESSAGES.get(registerResult);
     }
 }
